backend/requirements.txt
psycopg2-binary==2.9.9
asyncpg==0.30.0
Django==5.1.4
djangorestframework==3.14.0
markdown==3.7
django-filter==24.3
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
drf-spectacular==0.27.1
django-cors-headers==4.3.1